requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
msal
requests
schedule
python-dotenv
python-docx
pandas
openpyxl
